+++ lib/camera/camera.dart
@@ -44,6 +44,9 @@ class Camera with BonfireHasGameRef<BonfireGame> {
   double defaultShakeIntensity = 8.0; // in pixels
   double defaultShakeDuration = 1; // in seconds
 
+  double get zoom => config.zoom;
+  double get angle => config.angle;
+
   void moveTop(double displacement) {
     position = position.translate(0, displacement * -1);
   }
@@ -63,6 +66,7 @@ class Camera with BonfireHasGameRef<BonfireGame> {
   void moveToPositionAnimated(
     Offset position, {
     double zoom = 1,
+    double angle = 0,
     VoidCallback? finish,
     Duration? duration,
     Curve curve = Curves.decelerate,
@@ -79,6 +83,9 @@ class Camera with BonfireHasGameRef<BonfireGame> {
     double diffZoom = config.zoom - zoom;
     double initialZoom = config.zoom;
 
+    double diffAngle = config.angle - angle;
+    double originAngle = config.angle;
+
     gameRef.getValueGenerator(
       duration ?? Duration(seconds: 1),
       onChange: (value) {
@@ -89,6 +96,7 @@ class Camera with BonfireHasGameRef<BonfireGame> {
               y: originY - (diffY * value),
             );
         config.zoom = initialZoom - (diffZoom * value);
+        config.angle = originAngle - (diffAngle * value);
 
         if (config.moveOnlyMapArea) {
           _keepInMapArea();
@@ -105,6 +113,7 @@ class Camera with BonfireHasGameRef<BonfireGame> {
   void moveToTargetAnimated(
     GameComponent target, {
     double zoom = 1,
+    double angle = 0,
     VoidCallback? finish,
     Duration? duration,
     Curve curve = Curves.decelerate,
@@ -119,6 +128,9 @@ class Camera with BonfireHasGameRef<BonfireGame> {
     double diffZoom = config.zoom - zoom;
     double initialZoom = config.zoom;
 
+    double diffAngle = config.angle - angle;
+    double originAngle = config.angle;
+
     gameRef.getValueGenerator(
       duration ?? Duration(seconds: 1),
       onChange: (value) {
@@ -128,6 +140,7 @@ class Camera with BonfireHasGameRef<BonfireGame> {
         this.position = position.copyWith(x: originX - (diffX * value));
         this.position = position.copyWith(y: originY - (diffY * value));
         config.zoom = initialZoom - (diffZoom * value);
+        config.angle = originAngle - (diffAngle * value);
 
         if (config.moveOnlyMapArea) {
           _keepInMapArea();
@@ -159,12 +172,14 @@ class Camera with BonfireHasGameRef<BonfireGame> {
     Duration? duration,
     VoidCallback? finish,
     double zoom = 1,
+    double angle = 0,
     Curve curve = Curves.decelerate,
   }) {
     if (gameRef.player == null) return;
     moveToTargetAnimated(
       gameRef.player!,
       zoom: zoom,
+      angle: angle,
       finish: finish,
       duration: duration,
       curve: curve,
@@ -256,6 +271,30 @@ class Camera with BonfireHasGameRef<BonfireGame> {
       duration ?? Duration(seconds: 1),
       onChange: (value) {
         config.zoom = initialZoom - (diffZoom * value);
+      },
+      onFinish: () {
+        _isMoving = false;
+        finish?.call();
+      },
+      curve: curve,
+    ).start();
+  }
+
+  void animateRotate({
+    required double angle,
+    Duration? duration,
+    VoidCallback? finish,
+    Curve curve = Curves.decelerate,
+  }) {
+    _isMoving = true;
+
+    final diffAngle = config.angle - angle;
+    final originAngle = config.angle;
+
+    gameRef.getValueGenerator(
+      duration ?? Duration(seconds: 1),
+      onChange: (value) {
+        config.angle = originAngle - (diffAngle * value);
       },
       onFinish: () {
         _isMoving = false;