+++ lib/util/talk/talk_dialog.dart
@@ -8,17 +8,19 @@ class TalkDialog extends StatefulWidget {
   const TalkDialog({
     Key? key,
     required this.says,
-    this.finish,
+    this.onFinish,
     this.onChangeTalk,
     this.textBoxMinHeight = 100,
     this.keyboardKeysToNext,
     this.padding,
+    this.onClose,
   }) : super(key: key);
 
   static show(
     BuildContext context,
     List<Say> sayList, {
-    VoidCallback? finish,
+    VoidCallback? onFinish,
+    VoidCallback? onClose,
     ValueChanged<int>? onChangeTalk,
     Color? backgroundColor,
     double boxTextHeight = 100,
@@ -31,7 +33,8 @@ class TalkDialog extends StatefulWidget {
       builder: (BuildContext context) {
         return TalkDialog(
           says: sayList,
-          finish: finish,
+          onFinish: onFinish,
+          onClose: onClose,
           onChangeTalk: onChangeTalk,
           textBoxMinHeight: boxTextHeight,
           keyboardKeysToNext: logicalKeyboardKeysToNext,
@@ -42,7 +45,8 @@ class TalkDialog extends StatefulWidget {
   }
 
   final List<Say> says;
-  final VoidCallback? finish;
+  final VoidCallback? onFinish;
+  final VoidCallback? onClose;
   final ValueChanged<int>? onChangeTalk;
   final double? textBoxMinHeight;
   final List<LogicalKeyboardKey>? keyboardKeysToNext;
@@ -73,6 +77,7 @@ class _TalkDialogState extends State<TalkDialog> {
 
   @override
   void dispose() {
+    widget.onClose?.call();
     _textShowController.close();
     _focusNode.dispose();
     super.dispose();
@@ -186,7 +191,7 @@ class _TalkDialogState extends State<TalkDialog> {
       if (widget.onChangeTalk != null)
         widget.onChangeTalk?.call(currentIndexTalk);
     } else {
-      widget.finish?.call();
+      widget.onFinish?.call();
       Navigator.pop(context);
     }
   }
@@ -204,12 +209,14 @@ class _TalkDialogState extends State<TalkDialog> {
     _textShowController.add([TextSpan()]);
 
     await Future.forEach<TextSpan>(currentSay.text, (span) async {
+      if (_textShowController.isClosed) return;
       for (int i = 0; i < (span.text?.length ?? 0); i++) {
         if (finishedCurrentSay) {
           _textShowController.add([...currentSay.text]);
           break;
         }
         await Future.delayed(Duration(milliseconds: currentSay.speed ?? 50));
+        if (_textShowController.isClosed) return;
         _textShowController.add([
           ...currentSay.text.sublist(0, currentSay.text.indexOf(span)),
           TextSpan(