+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.9.7
+- Improvements in `TalkDialog.show` : add `onClose`.
+- Add rotation functionality to the camera. Set `angle` on `CameraConfig` or `animateRotate` to rotate the camera view
+
+
 # 1.9.6
 - Fix bug "getting stuck" in `MoveToPositionAlongThePath`
 - [BREAKING CHANGE] Change param `logicalKeyboardKeyToNext` in `TalkDialog` to `logicalKeyboardKeysToNext`, now multiple keys are accepted to advance in the dialogue